session: split hook entry, plugin-mode gates
